Use size_t indices in CPDF_TextPage.
Use std::reverse() and/or iterators to avoid reworking some
signed int comparisons.

Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/91511
Commit-Queue: Tom Sepez <tsepez@chromium.org>

+++ core/fpdftext/cpdf_textpage.h
@@ -130,7 +130,7 @@ class CPDF_TextPage {
       const CPDF_TextObject* pTextObj) const;
   TextOrientation FindTextlineFlowOrientation() const;
   void AppendGeneratedCharacter(wchar_t unicode, const CFX_Matrix& formMatrix);
-  void SwapTempTextBuf(int iCharListStartAppend, int iBufStartAppend);
+  void SwapTempTextBuf(size_t iCharListStartAppend, size_t iBufStartAppend);
   WideString GetTextByPredicate(
       const std::function<bool(const CharInfo&)>& predicate) const;
